Finally make the large data transfer reliable

+++ inc/timer.h
@@ -2,11 +2,12 @@
 #define _TIMER_H_
 
 #include <pthread.h>
+#include <stdint.h>
 #include <time.h>
 
 typedef struct {
     struct timespec ts;
-    int seq;
+    uint32_t seq;
     int len;
 } ts_pair;
 
@@ -20,7 +21,7 @@ typedef struct {
     long est_rtt; // the RTT in micro seconds, scaled
     long deviation; // in micro seconds, scaled
     long timeout; // in micro seconds
-    long largest_seq;
+    uint32_t largest_seq;
 
 	pthread_mutex_t lock;
 
@@ -29,8 +30,8 @@ typedef struct {
 tcp_timer_t *new_tcp_timer(int initial_rtt_msec, int cap);
 void tcp_timer_free(tcp_timer_t *timer);
 
-void timer_start_track(tcp_timer_t *timer, int seq, int len);
-int timer_end_track(tcp_timer_t *timer, int ack);
+void timer_start_track(tcp_timer_t *timer, uint32_t seq, int len);
+int timer_end_track(tcp_timer_t *timer, uint32_t ack);
 long timer_get_timeout(tcp_timer_t *timer);
 
 int get_curusec(struct timespec *ts);
